Wasm support in kube testing
