+++ packages/gravity/kube-testing/package.json
@@ -43,6 +43,7 @@
     "danfojs-node": "^1.1.2",
     "debug": "^4.3.4",
     "esbuild": "^0.19.6",
+    "esbuild-plugin-wasm": "^1.1.0",
     "ioredis": "^5.3.2",
     "isomorphic-ws": "^4.0.1",
     "js-yaml": "^4.1.0",
